pyproject.toml
[project]
name = "qualtrics-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "jinja2>=3.1.6",
    "playwright>=1.58.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "rapidfuzz>=3.14.5",
    "uvicorn>=0.44.0",
]
